requirements.txt
Django==3.0.5
django-filter==2.3.0
djangorestframework==3.11.1
djongo==1.3.3
pymongo==3.11.0
pytz==2020.1
sqlparse==0.2.4
